Update italics regex, again.
This time, our regex was cutting off whitespace before the italicized
word or phrase, as well as ignoring punctuation characters after it,
such as `:`.

This change ensures that (1) there is whitespace before the first
italicized word, and (2) we ignore all of the post-italic matching we
were trying to do.  I'm not even sure how that patch seemed like a good
idea, but it's just about time to drop everything and create a real test
suite in preparation for a big refactoring.