Fix bug: anchors weren't being added to headers
This isn't exactly a "fix".  A while back we added some logic to check
if the user had specified the TOC macro, and if they hadn't, to avoid
bothering to add anchor tags to headers.

Naturally this introduced a bug where NO headers were having anchors
added.

This commit adds a "quickfix" that just adds the headers back in no
matter what.  In future we should rip out that logic altogether since
it's not being used.