Fix bug with 'include' macro, add doc comments
By moving the block of code where we implement 'include' up in the
"waterfall" of the main DO block, we slurp in all of that pesky wikitext
early enough that the rest of our transformation rules get to run on it.

Of note is that, for now, we only support paragraphs with some inline
links/formatting, not bulleted lists and so on.  This is probably a bug,
but for now let's call it "how it works".